# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.AnswerRequestDTO:
    properties:
      answer:
        type: string
      questionId:
        type: integer
    required:
    - answer
    - questionId
    type: object
  dto.LoginDTO:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.QuestionRequestDTO:
    properties:
      question:
        type: string
    required:
    - question
    type: object
  helper.Response:
    properties:
      coce:
        type: integer
      data: {}
      error: {}
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /answer:
    get:
      consumes:
      - application/json
      description: Answer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - Bearer: []
      - JWT: []
      summary: Answer
      tags:
      - answer
    post:
      consumes:
      - application/json
      description: Answer Insert
      parameters:
      - description: request body insert question
        in: body
        name: register
        required: true
        schema:
          $ref: '#/definitions/dto.AnswerRequestDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - Bearer: []
      - JWT: []
      summary: Answer Insert
      tags:
      - answer
  /answer/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Answer By ID
      parameters:
      - description: Id Answer
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - Bearer: []
      - JWT: []
      summary: Delete Answer By ID
      tags:
      - answer
    get:
      consumes:
      - application/json
      description: Answer By ID
      parameters:
      - description: Id Answer
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - Bearer: []
      - JWT: []
      summary: Answer By ID
      tags:
      - answer
    put:
      consumes:
      - application/json
      description: Update Answwer By ID
      parameters:
      - description: Id Answer
        in: path
        name: id
        required: true
        type: integer
      - description: request body insert answer
        in: body
        name: register
        required: true
        schema:
          $ref: '#/definitions/dto.QuestionRequestDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - Bearer: []
      - JWT: []
      summary: Update Answwer By ID
      tags:
      - answer
  /login:
    post:
      consumes:
      - application/json
      description: 'Login user email : admin@gmail.com password : "admin", email :
        admin2@gmail.com password : "admin"'
      parameters:
      - description: request body login
        in: body
        name: register
        required: true
        schema:
          $ref: '#/definitions/dto.LoginDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      summary: Login user
      tags:
      - auth
  /profile:
    get:
      consumes:
      - application/json
      description: Profile user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - Bearer: []
      - JWT: []
      summary: Profile user
      tags:
      - accounts
  /question:
    get:
      consumes:
      - application/json
      description: Question
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - Bearer: []
      - JWT: []
      summary: Question
      tags:
      - question
    post:
      consumes:
      - application/json
      description: Question Insert
      parameters:
      - description: request body insert question
        in: body
        name: register
        required: true
        schema:
          $ref: '#/definitions/dto.QuestionRequestDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - Bearer: []
      - JWT: []
      summary: Question Insert
      tags:
      - question
  /question/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Question By ID
      parameters:
      - description: Id Question
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - Bearer: []
      - JWT: []
      summary: Delete Question By ID
      tags:
      - question
    get:
      consumes:
      - application/json
      description: Question By ID
      parameters:
      - description: Id Question
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - Bearer: []
      - JWT: []
      summary: Question By ID
      tags:
      - question
    put:
      consumes:
      - application/json
      description: Update Question By ID
      parameters:
      - description: Id Question
        in: path
        name: id
        required: true
        type: integer
      - description: request body insert question
        in: body
        name: register
        required: true
        schema:
          $ref: '#/definitions/dto.QuestionRequestDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - Bearer: []
      - JWT: []
      summary: Update Question By ID
      tags:
      - question
  /question/answer/{id}:
    get:
      consumes:
      - application/json
      description: Get Answer By ID Question
      parameters:
      - description: Id Question
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - Bearer: []
      - JWT: []
      summary: Get Answer ID Question
      tags:
      - question
securityDefinitions:
  JWT:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"